refactor: /home -> /
