+++ src/app/(subPages)/tour/templates/priceList.tsx
