+++ src/app/templates/homePhotoGallery.tsx
