+++ src/app/templates/templates.module.scss
@@ -10,7 +10,7 @@
 .main {
   width: 100%;
   height: 100dvh;
-  background-image: url('../../../../../public/home/top.jpg');
+  background-image: url('../../../public/home/top.jpg');
   background-repeat: no-repeat;
   background-size: cover;
   animation: fadeIn 2s ease-in;
